+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add tests for ResourceEditCustomTitle component. (UIEH-1095)
 * Add tests for TitleShow component. (UIEH-1098)
 * Add tests for AgreementsList component. (UIEH-1055)
+* Choose custom package - batch load packages. (UIEH-962)
 
 ## [6.1.0] (https://github.com/folio-org/ui-eholdings/tree/v6.1.0) (2021-06-04)
 * Add tests coverage for keyboard shortcuts to eholdings. (UIEH-1043)